# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Clipaso API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/clips/{clip_id}": {
      "patch": {
        "tags": [
          "clips"
        ],
        "summary": "Update Clip",
        "description": "Edita el título y los textos para publicar (no hace falta volver a renderizar).",
        "operationId": "update_clip_clips__clip_id__patch",
        "parameters": [
          {
            "name": "clip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Clip Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClipUpdateIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClipOut"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/clips/{clip_id}/rating": {
      "put": {
        "tags": [
          "clips"
        ],
        "summary": "Rate Clip",
        "operationId": "rate_clip_clips__clip_id__rating_put",
        "parameters": [
          {
            "name": "clip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Clip Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RatingIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClipOut"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/clips/{clip_id}/captions": {
      "get": {
        "tags": [
          "clips"
        ],
        "summary": "Clip Captions",
        "description": "Subtítulos del clip (con las correcciones del usuario) para editarlos en CapCut, Premiere…",
        "operationId": "clip_captions_clips__clip_id__captions_get",
        "parameters": [
          {
            "name": "clip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Clip Id"
            }
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "srt",
                "vtt"
              ],
              "type": "string",
              "default": "srt",
              "title": "Format"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/clips/{clip_id}/editor": {
      "get": {
        "tags": [
          "clips"
        ],
        "summary": "Clip Editor",
        "description": "Todo lo que necesita el editor: tramo del original con contexto, palabras y estilo.",
        "operationId": "clip_editor_clips__clip_id__editor_get",
        "parameters": [
          {
            "name": "clip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Clip Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EditorOut"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/clips/{clip_id}/render": {
      "post": {
        "tags": [
          "clips"
        ],
        "summary": "Render Clip",
        "description": "Guarda las ediciones (recorte, palabras, estilo) y vuelve a generar el clip.",
        "operationId": "render_clip_clips__clip_id__render_post",
        "parameters": [
          {
            "name": "clip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Clip Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenderIn"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClipOut"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/clips/{clip_id}/exports": {
      "get": {
        "tags": [
          "clips"
        ],
        "summary": "Clip Exports",
        "description": "Descargas disponibles del clip: MP4 en varias calidades y el audio en MP3.",
        "operationId": "clip_exports_clips__clip_id__exports_get",
        "parameters": [
          {
            "name": "clip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Clip Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExportsOut"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "clips"
        ],
        "summary": "Request Export",
        "description": "Genera una descarga: el MP3 al momento; otra calidad de vídeo, en segundo plano.",
        "operationId": "request_export_clips__clip_id__exports_post",
        "parameters": [
          {
            "name": "clip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Clip Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExportsOut"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ClipOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "rank": {
            "type": "integer",
            "title": "Rank"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "hashtags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Hashtags"
          },
          "rating": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rating"
          },
          "start": {
            "type": "number",
            "title": "Start"
          },
          "end": {
            "type": "number",
            "title": "End"
          },
          "duration": {
            "type": "number",
            "title": "Duration"
          },
          "score": {
            "type": "number",
            "title": "Score"
          },
          "status": {
            "type": "string",
            "enum": [
              "ready",
              "rendering",
              "failed"
            ],
            "title": "Status"
          },
          "render_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Render Error"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "video_url": {
            "type": "string",
            "title": "Video Url"
          },
          "download_url": {
            "type": "string",
            "title": "Download Url"
          },
          "thumbnail_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thumbnail Url"
          }
        },
        "type": "object",
        "required": [
          "id",
          "rank",
          "title",
          "reason",
          "description",
          "hashtags",
          "rating",
          "start",
          "end",
          "duration",
          "score",
          "status",
          "render_error",
          "version",
          "video_url",
          "download_url",
          "thumbnail_url"
        ],
        "title": "ClipOut"
      },
      "ClipUpdateIn": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "hashtags": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array",
                "maxItems": 15
              },
              {
                "type": "null"
              }
            ],
            "title": "Hashtags"
          }
        },
        "type": "object",
        "title": "ClipUpdateIn"
      },
...
      "EditorOut": {
        "properties": {
          "clip": {
            "$ref": "#/components/schemas/ClipOut"
          },
          "project_title": {
            "type": "string",
            "title": "Project Title"
          },
          "format": {
            "type": "string",
            "enum": [
              "vertical",
              "square",
              "horizontal"
            ],
            "title": "Format",
            "description": "Encuadre del clip (en el formato «original», el del vídeo)."
          },
          "source_duration": {
            "type": "number",
            "title": "Source Duration"
          },
          "window_start": {
            "type": "number",
            "title": "Window Start"
          },
          "window_end": {
            "type": "number",
            "title": "Window End"
          },
          "preview_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preview Url",
            "description": "Vídeo ligero del original para previsualizar el corte."
          },
          "words": {
            "items": {
              "$ref": "#/components/schemas/EditorWordOut"
            },
            "type": "array",
            "title": "Words"
          },
          "caption_style": {
            "$ref": "#/components/schemas/CaptionStyle"
          },
          "can_render": {
            "type": "boolean",
            "title": "Can Render"
          },
          "energy": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Energy",
            "description": "Volumen del audio (0-1) desde window_start."
          },
          "energy_step": {
            "type": "number",
            "title": "Energy Step",
            "description": "Segundos entre valores de `energy`.",
            "default": 0.5
          }
        },
        "type": "object",
        "required": [
          "clip",
          "project_title",
          "format",
          "source_duration",
          "window_start",
          "window_end",
          "preview_url",
          "words",
          "caption_style",
          "can_render"
        ],
        "title": "EditorOut"
      },
      "EditorWordOut": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key"
          },
          "start": {
            "type": "number",
            "title": "Start"
          },
          "end": {
            "type": "number",
            "title": "End"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "original": {
            "type": "string",
            "title": "Original"
          },
          "brk": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "split",
                  "join"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Brk",
            "description": "Corte de línea elegido tras esta palabra."
          }
        },
        "type": "object",
        "required": [
          "key",
          "start",
          "end",
          "text",
          "original"
        ],
        "title": "EditorWordOut"
      },
      "ErrorBody": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse"
      },
      "ExportIn": {
        "properties": {
          "format": {
            "type": "string",
            "enum": [
              "mp4",
              "mp3"
            ],
            "title": "Format"
          },
          "quality": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "480p",
                  "720p",
                  "1080p",
                  "2160p"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Quality"
          }
        },
        "type": "object",
        "required": [
          "format"
        ],
        "title": "ExportIn"
      },
      "ExportOut": {
        "properties": {
          "format": {
            "type": "string",
            "enum": [
              "mp4",
              "mp3"
            ],
            "title": "Format"
          },
          "quality": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "480p",
                  "720p",
                  "1080p",
                  "2160p"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Quality"
          },
          "status": {
            "type": "string",
            "enum": [
              "ready",
              "pending",
              "failed",
              "available"
            ],
            "title": "Status",
            "description": "ready: se puede descargar · pending: generándose · available: se puede pedir"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          },
          "size_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Size Bytes"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          }
        },
        "type": "object",
        "required": [
          "format",
          "quality",
          "status"
        ],
        "title": "ExportOut"
      },
      "ExportsOut": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ExportOut"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "ExportsOut"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RatingIn": {
        "properties": {
          "value": {
            "type": "integer",
            "enum": [
              -1,
              0,
              1
            ],
            "title": "Value"
          }
        },
        "type": "object",
        "required": [
          "value"
        ],
        "title": "RatingIn"
      },
      "RenderIn": {
        "properties": {
          "start": {
            "type": "number",
            "minimum": 0.0,
            "title": "Start"
          },
          "end": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "End"
          },
          "word_edits": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "maxProperties": 5000,
            "title": "Word Edits"
          },
          "caption_style": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CaptionStyle"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "start",
          "end"
        ],
        "title": "RenderIn"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}